requirement.txt
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.57.0
accelerate>=0.26.0
huggingface-hub>=0.23.0
safetensors>=0.4.3
pillow>=10.0.0
matplotlib>=3.8.0
sentencepiece>=0.2.0
protobuf>=4.25.0
qwen-vl-utils>=0.0.8
